completed

+++ ConsoleApp13/Controllers/GroupController.cs
@@ -20,26 +20,36 @@ namespace Manage.Controllers
         #region CreateGroup
         public void CreateGroup()
         {
-            ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkMagenta, "Enter Group Name");
+        Name: ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkMagenta, "Enter Group Name");
             string name = Console.ReadLine();
-        MaxSize: ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkMagenta, "Enter Group maxSize");
-            string size = Console.ReadLine();
-            int maxSize;
-            bool result = int.TryParse(size, out maxSize);
-            if (result)
+            var group = _groupRepository.Get(g => g.Name.ToLower() == name.ToLower());
+            if (group == null)
             {
-                Group group = new Group
+            MaxSize: ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkMagenta, "Enter Group maxSize");
+                string size = Console.ReadLine();
+                int maxSize;
+                bool result = int.TryParse(size, out maxSize);
+                if (result)
                 {
-                    Name = name,
-                    MaxSize = maxSize
-                };
-                var createdGroup = _groupRepository.Create(group);
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.Green, $"{createdGroup.Name} is succesfully created with maxSize -{createdGroup.MaxSize}");
+                    Group newGroup = new Group
+                    {
+                        Name = name,
+                        MaxSize = maxSize
+                    };
+                    var createdGroup = _groupRepository.Create(newGroup);
+                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Green, $"{createdGroup.Name} is succesfully created with maxSize -{createdGroup.MaxSize}");
+                }
+                else
+                {
+                    ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkRed, "Please Enter Correct Group maxSize");
+                    goto MaxSize;
+                }
+
             }
             else
             {
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkRed, "Please Enter Correct Group maxSize");
-                goto MaxSize;
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkRed, "This group already exist");
+                goto Name;
             }
         }
         #endregion
@@ -48,40 +58,67 @@ namespace Manage.Controllers
 
         public void UpdateGroup()
         {
-            ConsoleHelper.WriteTextWithColor(ConsoleColor.Magenta, "Enter Group Name:");
-            string name = Console.ReadLine();
-            var group = _groupRepository.Get(g => g.Name.ToLower() == name.ToLower());
-            if (group != null)
+            var groups = _groupRepository.GetAll();
+            if (groups.Count > 0)
             {
-                int oldSize = group.MaxSize;
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkCyan, "Enter new group name ");
-                string newName = Console.ReadLine();
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.Yellow, "All groups");
+                foreach (var dbGroup in groups)
+                {
+                    Console.WriteLine($"Id: {dbGroup.Id}, Name:{dbGroup.Name}");
+                }
 
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkCyan, " Enter new group max size");
-                string size = Console.ReadLine();
-
-                int maxSize;
-                bool result = int.TryParse(size, out maxSize);
+            Id: ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkCyan, "Enter group id:");
+                int chosenId;
+                string id = Console.ReadLine();
+                var result = int.TryParse(id, out chosenId);
                 if (result)
                 {
-                    var newGroup = new Group
+                    var group = _groupRepository.Get(g => g.Id == chosenId);
+                    if (group != null)
                     {
-                        Id = group.Id,
-                        Name = newName,
-                        MaxSize = maxSize
-                    };
-                    _groupRepository.Update(newGroup);
+                        int oldSize = group.MaxSize;
+                        string oldName = group.Name;
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkCyan, "Enter new group name ");
+                        string newName = Console.ReadLine();
 
-                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Green, $"Name:{name} Max size: {oldSize} is updated to Name: {newName} Max size: {maxSize} ");
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkCyan, " Enter new group max size");
+                        string size = Console.ReadLine();
+
+                        int maxSize;
+                        result = int.TryParse(size, out maxSize);
+
+                        if (result)
+                        {
+                            var newGroup = new Group
+                            {
+                                Id = group.Id,
+                                Name = newName,
+                                MaxSize = maxSize
+                            };
+                            _groupRepository.Update(newGroup);
+
+                            ConsoleHelper.WriteTextWithColor(ConsoleColor.Green, $"Name:{oldName} Max size: {oldSize} is updated to Name: {newGroup.Name} Max size: {newGroup.MaxSize} ");
+                        }
+                        else
+                        {
+                            ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "Please, enter correct group max size:");
+                        }
+                    }
+                    else
+                    {
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "Please, enter correct group name:");
+                    }
+
                 }
                 else
                 {
-                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "Please, enter correct group max size:");
+                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "Please, enter correct id:");
+                    goto Id;
                 }
             }
             else
             {
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "Please, enter correct group name:");
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "There are no any groups");
             }
         }
 
@@ -90,21 +127,45 @@ namespace Manage.Controllers
         #region DeleteGroup
         public void DeleteGroup()
         {
-            ConsoleHelper.WriteTextWithColor(ConsoleColor.Yellow, "Enter Group Name:");
-            string name = Console.ReadLine();
-            var group = _groupRepository.Get(g => g.Name.ToLower() == name.ToLower());
-
-            if (group != null)
+            var groups = _groupRepository.GetAll();
+            if (groups.Count > 0)
             {
-                _groupRepository.Delete(group);
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.Green, $"{name} is deleted");
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.Yellow, "All Groups");
+                foreach (var dbGroup in groups)
+                {
+                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Magenta, $"Id: {dbGroup.Id}, Name: {dbGroup.Name}");
+                }
+
+            Id: ConsoleHelper.WriteTextWithColor(ConsoleColor.Yellow, "Enter Group Id:");
+                int chosenId;
+                string id = Console.ReadLine();
+                var result = int.TryParse(id, out chosenId);
+                if (result)
+                {
+                    var group = _groupRepository.Get(g => g.Id == chosenId);
+                    if (group != null)
+                    {
+                        string name = group.Name;
+                        _groupRepository.Delete(group);
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.Green, $"{name} is deleted");
+                    }
+                    else
+                    {
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "This group doesnt exist:");
+                    }
+
+                }
+                else
+                {
+                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "Please enter correct group Id");
+                    goto Id;
+                }
             }
             else
             {
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "This Group Doesnt Exist:");
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "There are no any groups");
             }
         }
-
         #endregion
 
         #region AllGroups
@@ -114,7 +175,7 @@ namespace Manage.Controllers
             ConsoleHelper.WriteTextWithColor(ConsoleColor.Magenta, "All Groups");
             foreach (var group in groups)
             {
-                Console.WriteLine($"{group.Name}, {group.MaxSize}");
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.Cyan,$"Name:{group.Name}, MaxSize:{group.MaxSize}");
             }
         }
         #endregion
@@ -141,7 +202,7 @@ namespace Manage.Controllers
         #region Exit
         public void Exit()
         {
-          ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkYellow, "Thanks for using our application");
+            ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkYellow, "Thanks for using our application");
         }
 
         #endregion
